fix: black bar on screen not clearing frames
Changed MainCamera's viewport rect Y calue from 0.1 to 0.
While it was 0.1, the camera size did not conform with the game
size, causing a bar at the bottom of the screen which which did
not clear on each frame, causing previous drawn frames to remain
on screen layering on top of each other.

Changing the viewport rect does slightly push the camera higher up,
so a camera offset may be good to add so the jump and scores do not
go over the player character